refactor: move dirs
