+++ src/api/admin/courses/router.py
@@ -5,8 +5,8 @@ from typing import TYPE_CHECKING
 from fastapi import APIRouter, Body, Depends, Path, status
 from fastapi.responses import JSONResponse
 
-from src.api.admin.dependencies import get_admin, get_admin_courses_query_service
-from src.api.admin.schemas import CreateCourseRequest, CreateCourseResponse, UpdateCourseRequest
+from src.api.admin.courses.dependencies import get_admin, get_admin_courses_query_service
+from src.api.admin.courses.schemas import CreateCourseRequest, CreateCourseResponse, UpdateCourseRequest
 from src.api.base_schemas import ErrorResponse, SuccessResponse
 from src.api.courses.dependencies import get_courses_command_service, get_talent_courses_query_service
 from src.api.courses.schemas import CourseFullDTO, CourseShortDTO
@@ -30,7 +30,7 @@ router = APIRouter(prefix="/admin", tags=["admin"])
 
 
 @router.post(
-    "/courses",
+    "",
     status_code=status.HTTP_201_CREATED,
     description="Create new course",
     summary="Create course",
@@ -258,7 +258,7 @@ async def get_course(
 
 
 @router.get(
-    "/courses",
+    "",
     status_code=status.HTTP_200_OK,
     description="Get all courses",
     summary="Get courses",
