+++ src/api/admin/courses/schemas.py
